fix: fixing according to the review
